simpleRentals/requirements.txt
asgiref==3.8.1
Django==5.1.6
pillow==11.1.0
psycopg==3.2.4
setuptools==75.8.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
wheel==0.45.1
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
psycopg2-binary
python-dotenv
pytest